% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateOrthologSimilarity.R
\name{calculateCompositeSimilarity}
\alias{calculateCompositeSimilarity}
\title{Calculate composite similarity score}
\usage{
calculateCompositeSimilarity(data, debug = FALSE)
}
\arguments{
\item{data}{Data frame with similarity scores}

\item{debug}{Debug mode}
}
\value{
Data frame with composite similarity score
}
\description{
Calculate composite similarity score
}
