% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateOrthologSimilarity.R,
%   R/patristicDistances.R
\name{calculateEvolutionarySimilarity}
\alias{calculateEvolutionarySimilarity}
\title{Calculate evolutionary similarity between orthologs}
\usage{
calculateEvolutionarySimilarity(patristic_distance, max_distance = 1)

calculateEvolutionarySimilarity(patristic_distance, max_distance = 1)
}
\arguments{
\item{patristic_distance}{Patristic distance in substitutions per site}

\item{max_distance}{Maximum expected distance for normalization (default: 1.0)}

\item{data}{Data frame with ortholog information}

\item{species1}{Source species}

\item{species2}{Target species}

\item{debug}{Debug mode}
}
\value{
Data frame with evolutionary similarity scores

Evolutionary similarity score (0-1)
}
\description{
Calculates evolutionary similarity score from patristic distance
}
