% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateOrthologSimilarity.R
\name{calculateFunctionalSimilarity}
\alias{calculateFunctionalSimilarity}
\title{Calculate functional similarity between orthologs}
\usage{
calculateFunctionalSimilarity(data, species1, species2, debug = FALSE)
}
\arguments{
\item{data}{Data frame with ortholog information}

\item{species1}{Source species}

\item{species2}{Target species}

\item{debug}{Debug mode}
}
\value{
Data frame with functional similarity scores
}
\description{
Calculate functional similarity between orthologs
}
