% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patristicDistances.R
\name{calculatePatristicDistance}
\alias{calculatePatristicDistance}
\title{Calculate patristic distance from phylogenetic tree}
\usage{
calculatePatristicDistance(tree, species1, species2, debug = FALSE)
}
\arguments{
\item{tree}{Phylogenetic tree object (ape::phylo)}

\item{species1}{First species name}

\item{species2}{Second species name}

\item{debug}{Debug mode for verbose output}
}
\value{
Patristic distance in substitutions per site
}
\description{
Calculates patristic distance between two species from a phylogenetic tree
}
