% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateOrthologSimilarity.R
\name{calculateSequenceSimilarity}
\alias{calculateSequenceSimilarity}
\title{Calculate sequence similarity between orthologs}
\usage{
calculateSequenceSimilarity(data, species1, species2, debug = FALSE)
}
\arguments{
\item{data}{Data frame with ortholog information}

\item{species1}{Source species}

\item{species2}{Target species}

\item{debug}{Debug mode}
}
\value{
Data frame with sequence similarity scores
}
\description{
Calculate sequence similarity between orthologs
}
