% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organismSupportTools.R
\name{checkDatabaseAvailability}
\alias{checkDatabaseAvailability}
\title{Check database availability for organism}
\usage{
checkDatabaseAvailability(orgm, verbose = FALSE)
}
\arguments{
\item{orgm}{A character string specifying the organism abbreviation}

\item{verbose}{A logical value indicating whether to print additional information. Defaults to FALSE.}
}
\value{
A list containing database availability information and installation guidance
}
\description{
This function checks if required databases are available for an organism
and provides installation guidance.
}
\examples{
\dontrun{
  # Check database availability for human
  db_status <- checkDatabaseAvailability("Hsapiens")
}

}
