% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organismValidation.R
\name{checkDatabaseAvailabilityInternal}
\alias{checkDatabaseAvailabilityInternal}
\title{Check database availability for organism (internal)}
\usage{
checkDatabaseAvailabilityInternal(orgm)
}
\arguments{
\item{orgm}{A character string specifying the organism abbreviation}
}
\value{
A list containing database availability information
}
\description{
This function checks if required databases are available for an organism.
This is the internal implementation used by other functions.
}
\examples{
\dontrun{
  # Check database availability for human
  db_status <- checkDatabaseAvailabilityInternal("Hsapiens")
}

}
