% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organismValidation.R
\name{checkPlottingCompatibility}
\alias{checkPlottingCompatibility}
\title{Check plotting compatibility for organism list}
\usage{
checkPlottingCompatibility(orgms_list, verbose = FALSE)
}
\arguments{
\item{orgms_list}{A character vector of organism abbreviations}

\item{verbose}{A logical value indicating whether to print additional information. Defaults to FALSE.}
}
\value{
A list containing compatibility status and details
}
\description{
This function validates plotting compatibility for a list of organisms.
}
\examples{
\dontrun{
  # Check plotting compatibility for multiple organisms
  compatibility <- checkPlottingCompatibility(c("Hsapiens", "Mmusculus", "Athaliana"))
}

}
