% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patristicDistances.R
\name{divergenceTimeToPatristic}
\alias{divergenceTimeToPatristic}
\title{Convert divergence time to patristic distance}
\usage{
divergenceTimeToPatristic(
  divergence_time_mya,
  mutation_rate = 1e-09,
  generation_time = 1
)
}
\arguments{
\item{divergence_time_mya}{Divergence time in millions of years}

\item{mutation_rate}{Mutation rate per site per year (default: 1e-9)}

\item{generation_time}{Generation time in years (default: 1)}
}
\value{
Patristic distance in substitutions per site
}
\description{
Converts divergence time to patristic distance using molecular clock
}
