% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSubset.R
\name{geneSubset}
\alias{geneSubset}
\title{Search and annotate genes against known gene databases based on AnnotationDbi}
\usage{
geneSubset(mycoords.gr, orgm)
}
\arguments{
\item{mycoords.gr}{A target search range in form of a GRanges object}

\item{orgm}{An organism that of interest. Options include:\cr
Btaurus(Bt) - Bos taurus - Cow \cr
Celegans(Ce) - Caenorhabditis elegans - Roundworm \cr
Cfamiliaris(Cf) - Canis familiaris - Dog \cr
Dmelanogaster(Dm) - Drosophila melanogaster - Fruitfly \cr
Drerio(Dr) - Danio rerio - Zebrafish \cr
Ggallus(Gg) - Gallus gallus - Chicken \cr
Hsapiens(Hs) - Homo Sapiens - Human \cr
Mmulatta(Mmu) - Macaca mulatta - Rhesus macaque \cr
Mmusculus(Mm) - Mus musculus - House mouse \cr
Ptroglodytes(Pt) - Pan troglodytes - Chimpanzee \cr
Rnorvegicus(Rn) - Rattus norvegicus - Brown rat \cr
Scerevisiae(Sc) - Saccharomyces cerevisiae - Brewer's yeast \cr
Sscrofa(Ss) - Sus scrofa - Wild swine}
}
\value{
A target search range as a GRanges object
}
\description{
Search and annotate genes against known gene databases based on AnnotationDbi
}
\examples{
\dontrun{
  mycoords.list <- "2:16e7:16.5e7"
  mycoords.gr <- coordFormat (mycoords.list)
  orgm <- "Hsapiens"
  geneSubset (mycoords.gr, orgm)
}
}
