% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinateUtils.R
\name{getGenomeAssembly}
\alias{getGenomeAssembly}
\title{Get genome assembly for species}
\usage{
getGenomeAssembly(species)
}
\arguments{
\item{species}{A character string specifying the species abbreviation}
}
\value{
A character string with the genome assembly name
}
\description{
This function returns the appropriate genome assembly for a given species.
}
\examples{
\dontrun{
  # Get genome assembly for human
  assembly <- getGenomeAssembly("Hsapiens")
}

}
