% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organismValidation.R
\name{getOrganismCapabilitiesInternal}
\alias{getOrganismCapabilitiesInternal}
\title{Get detailed organism capabilities (internal)}
\usage{
getOrganismCapabilitiesInternal(orgm, verbose = FALSE)
}
\arguments{
\item{orgm}{A character string specifying the organism abbreviation}

\item{verbose}{A logical value indicating whether to print additional information. Defaults to FALSE.}
}
\value{
A list containing detailed capability information
}
\description{
This function returns detailed capability information for an organism.
This is the internal implementation used by other functions.
}
\examples{
\dontrun{
  # Get detailed capabilities for human
  capabilities <- getOrganismCapabilitiesInternal("Hsapiens")
}

}
