% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDB.R
\name{getOrgDB}
\alias{getOrgDB}
\title{Search for orgDB}
\usage{
getOrgDB(orgm, orgOrgDB)
}
\arguments{
\item{orgm}{An organism that of interest. Options include:\cr
Btaurus(Bt) - Bos taurus - Cow \cr
Celegans(Ce) - Caenorhabditis elegans - Roundworm \cr
Cfamiliaris(Cf) - Canis familiaris - Dog \cr
Dmelanogaster(Dm) - Drosophila melanogaster - Fruitfly \cr
Drerio(Dr) - Danio rerio - Zebrafish \cr
Ggallus(Gg) - Gallus gallus - Chicken \cr
Hsapiens(Hs) - Homo Sapiens - Human \cr
Mmulatta(Mmu) - Macaca mulatta - Rhesus macaque \cr
Mmusculus(Mm) - Mus musculus - House mouse \cr
Ptroglodytes(Pt) - Pan troglodytes - Chimpanzee \cr
Rnorvegicus(Rn) - Rattus norvegicus - Brown rat \cr
Scerevisiae(Sc) - Saccharomyces cerevisiae - Brewer's yeast \cr
Sscrofa(Ss) - Sus scrofa - Wild swine}

\item{orgOrgDB}{A data.frame that contains available orgDB object, e.g. \code{\link{orgmOrgDB}}.}
}
\description{
Search if there is a corresponding annotation database orgDB is available for the given organism.
}
\section{getDB complementing functions}{

A collection of complementing functions that required by the main functions:
\code{getPkgs}
, which includes following functions:
\enumerate{
  \item{\code{getOrgDB}}
  \item{\code{getTxDB}}
}
}

\examples{
orgm <- "Hsapiens"
getOrgDB(orgm,orgmOrgDB)
}
\seealso{
Other orgmOrgDB: 
\code{\link{orgmOrgDB}}

Other getDB: 
\code{\link{getPkgs}()},
\code{\link{getTxDB}()}
}
\concept{getDB}
\concept{orgmOrgDB}
