% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOrthHomolog.R
\name{getOrthHomolog}
\alias{getOrthHomolog}
\title{Search for orthologous genes in a target species}
\usage{
getOrthHomolog(
  species,
  gene_id,
  gene_id_type = "ensembl_gene_id",
  verbose = FALSE,
  debug = FALSE
)
}
\arguments{
\item{species}{A character string specifying the target species (e.g., "mouse").}

\item{gene_id}{A character string specifying the gene ID (e.g., "ENSG00000139618").}

\item{gene_id_type}{A character string specifying the type of gene ID provided. Options are "ensembl_gene_id" or "symbol".}

\item{verbose}{A logical value indicating whether to print additional information during the function execution. Defaults to FALSE.}

\item{debug}{A logical value for additional debugging output. Defaults to FALSE.}
}
\value{
A data frame containing the orthologous gene(s) and their associated information.
}
\description{
This function uses the `orthogene` package to search for orthologous genes for a given target species.
}
\examples{
\dontrun{
 # Example usage of the getOrthHomolog function
 getOrthHomolog(species = "mouse", gene_id = "ENSG00000139618", gene_id_type = "ensembl_gene_id")
}

}
