% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patristicDistances.R
\name{getPatristicDistanceSummary}
\alias{getPatristicDistanceSummary}
\title{Summary statistics for patristic distances}
\usage{
getPatristicDistanceSummary()
}
\value{
List with summary statistics
}
\description{
Provides summary statistics for all available patristic distances
}
