% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDB.R
\name{getPkgs}
\alias{getPkgs}
\title{Download organism AnnotationDbi}
\usage{
getPkgs(orgm, orgDB)
}
\arguments{
\item{orgm}{An organism that of interest. Options include:\cr
Btaurus(Bt) - Bos taurus - Cow \cr
Celegans(Ce) - Caenorhabditis elegans - Roundworm \cr
Cfamiliaris(Cf) - Canis familiaris - Dog \cr
Dmelanogaster(Dm) - Drosophila melanogaster - Fruitfly \cr
Drerio(Dr) - Danio rerio - Zebrafish \cr
Ggallus(Gg) - Gallus gallus - Chicken \cr
Hsapiens(Hs) - Homo Sapiens - Human \cr
Mmulatta(Mmu) - Macaca mulatta - Rhesus macaque \cr
Mmusculus(Mm) - Mus musculus - House mouse \cr
Ptroglodytes(Pt) - Pan troglodytes - Chimpanzee \cr
Rnorvegicus(Rn) - Rattus norvegicus - Brown rat \cr
Scerevisiae(Sc) - Saccharomyces cerevisiae - Brewer's yeast \cr
Sscrofa(Ss) - Sus scrofa - Wild swine}

\item{orgDB}{A data.frame collection that contains available AnnotationDbi objs.
e.g. \code{\link{orgmTxDB}} or \code{\link{orgmOrgDB}}.}
}
\description{
Download required AnnotationDbi class object if there is a corresponding database is available
for the given organism.
}
\section{getDB main function}{

main function includes: \code{getPkgs}
}

\examples{
\dontrun{
  orgm <- "Hsapiens"
  getPkgs (orgm, orgmTxDB)
}
}
\seealso{
Other getDB: 
\code{\link{getOrgDB}()},
\code{\link{getTxDB}()}
}
\concept{getDB}
