% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organismValidation.R
\name{getSupportedOperations}
\alias{getSupportedOperations}
\title{Get supported operations for an organism}
\usage{
getSupportedOperations(orgm, verbose = FALSE)
}
\arguments{
\item{orgm}{A character string specifying the organism abbreviation}

\item{verbose}{A logical value indicating whether to print additional information. Defaults to FALSE.}
}
\value{
A list containing supported operations and capability details
}
\description{
This function returns a list of all supported operations for an organism.
}
\examples{
\dontrun{
  # Get all capabilities for human
  capabilities <- getSupportedOperations("Hsapiens")
}

}
