% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDB.R
\name{getTxDB}
\alias{getTxDB}
\title{Search for TxDB}
\usage{
getTxDB(orgm, orgTxDB)
}
\arguments{
\item{orgm}{An organism that of interest. Options include:\cr
Btaurus(Bt) - Bos taurus - Cow \cr
Celegans(Ce) - Caenorhabditis elegans - Roundworm \cr
Cfamiliaris(Cf) - Canis familiaris - Dog \cr
Dmelanogaster(Dm) - Drosophila melanogaster - Fruitfly \cr
Drerio(Dr) - Danio rerio - Zebrafish \cr
Ggallus(Gg) - Gallus gallus - Chicken \cr
Hsapiens(Hs) - Homo Sapiens - Human \cr
Mmulatta(Mmu) - Macaca mulatta - Rhesus macaque \cr
Mmusculus(Mm) - Mus musculus - House mouse \cr
Ptroglodytes(Pt) - Pan troglodytes - Chimpanzee \cr
Rnorvegicus(Rn) - Rattus norvegicus - Brown rat \cr
Scerevisiae(Sc) - Saccharomyces cerevisiae - Brewer's yeast \cr
Sscrofa(Ss) - Sus scrofa - Wild swine}

\item{orgTxDB}{A data.frame that contains currently TxDB objects, e.g. \code{\link{orgmTxDB}}.}
}
\description{
Search if there is a corresponding annotation database TxDB is available for the given organism.
}
\section{getDB complementing functions}{

A collection of complementing functions that required by the main functions:
\code{getPkgs}
, which includes following functions:
\enumerate{
  \item{\code{getOrgDB}}
  \item{\code{getTxDB}}
}
}

\examples{
orgm <- "Hsapiens"
getTxDB (orgm, orgmTxDB)
}
\seealso{
Other orgTxDB: 
\code{\link{orgmTxDB}}

Other getDB: 
\code{\link{getOrgDB}()},
\code{\link{getPkgs}()}
}
\concept{getDB}
\concept{orgTxDB}
