% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateOrthologSimilarity.R
\name{loadEvolutionaryDistances}
\alias{loadEvolutionaryDistances}
\title{Load evolutionary distances from YAML file}
\usage{
loadEvolutionaryDistances(debug = FALSE)
}
\arguments{
\item{debug}{Debug mode}
}
\value{
List of evolutionary distances
}
\description{
Load evolutionary distances from YAML file
}
\keyword{internal}
