% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patristicDistances.R
\name{loadPatristicDistances}
\alias{loadPatristicDistances}
\title{Load patristic distances from YAML file}
\usage{
loadPatristicDistances(debug = FALSE)
}
\arguments{
\item{debug}{Debug mode for verbose output}
}
\value{
List of patristic distances between species pairs
}
\description{
Loads patristic distances based on published phylogenetic studies
}
