% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synvizPlot.R
\name{multisynvizPlots}
\alias{multisynvizPlots}
\title{multisynvizPlots
A multi synteny plot for comparison and new insight discovery}
\usage{
multisynvizPlots(orgmsCollection, verbose = FALSE)
}
\arguments{
\item{orgmsCollection}{A GRangesList object that contains a list of search targets in GRange form}
}
\value{
A multi synteny plot.
}
\description{
multisynvizPlots
A multi synteny plot for comparison and new insight discovery
}
\examples{
\donttest{
  orgmsList <- orgmsCollection.init (orgmsList)
  orgm <- "Hsapiens"
  mycoords.list <- "2:16e7:16.5e7"
  orgmsList <- orgmsAdd (orgm, orgmTxDB, mycoords.list, orgmsList)
  multisynvizPlots(orgmsList)
}
}
\seealso{
Other SynvizPlot: 
\code{\link{synvizDataTest}},
\code{\link{synvizPlot}()},
\code{\link{synvizPlotData}()}
}
\concept{SynvizPlot}
