% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgmOrgDB-data.R
\docType{data}
\name{orgmOrgDB}
\alias{orgmOrgDB}
\title{A list of available orgDB objects as a data.frame}
\format{
A data.frame in table format
}
\source{
Currently available orgDB objects from
\href{http://bioconductor.org/packages/release/BiocViews.html#___AnnotationData}{bioconductor}
}
\usage{
data(orgmOrgDB)
}
\description{
orgmOrgDB is a data.frame that contains currently available orgDB objects from bioconductor
}
\examples{
data(orgmOrgDB)
databaseClass <- orgmOrgDB$dbClass
databaseSources <- orgmOrgDB$dbSource
species <- orgmOrgDB$dbSpecies
databaseTypes <- orgmOrgDB$dbAbbv
}
\seealso{
Other orgmOrgDB: 
\code{\link{getOrgDB}()}
}
\concept{orgmOrgDB}
\keyword{datasets}
