% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgmsCollection.R
\name{orgmsAdd}
\alias{orgmsAdd}
\title{orgmsAdd
Add a new search target in GRange format into GRangeList}
\usage{
orgmsAdd(orgm, mycoords.list, orgmsCollection)
}
\arguments{
\item{orgm}{An organism that of interest. Options include:\cr
Btaurus(Bt) - Bos taurus - Cow \cr
Celegans(Ce) - Caenorhabditis elegans - Roundworm \cr
Cfamiliaris(Cf) - Canis familiaris - Dog \cr
Dmelanogaster(Dm) - Drosophila melanogaster - Fruitfly \cr
Drerio(Dr) - Danio rerio - Zebrafish \cr
Ggallus(Gg) - Gallus gallus - Chicken \cr
Hsapiens(Hs) - Homo Sapiens - Human \cr
Mmulatta(Mmu) - Macaca mulatta - Rhesus macaque \cr
Mmusculus(Mm) - Mus musculus - House mouse \cr
Ptroglodytes(Pt) - Pan troglodytes - Chimpanzee \cr
Rnorvegicus(Rn) - Rattus norvegicus - Brown rat \cr
Scerevisiae(Sc) - Saccharomyces cerevisiae - Brewer's yeast \cr
Sscrofa(Ss) - Sus scrofa - Wild swine}

\item{orgmsCollection}{A GRangesList object that contains a list of search targets in GRange form}
}
\value{
orgmsCollection A GRangesList of search targets in GRange form
}
\description{
orgmsAdd
Add a new search target in GRange format into GRangeList
}
\examples{
orgm <- "Hsapiens"
mycoords.list <- "2:16e7:16.5e7"

}
\seealso{
Other orgmsCollection: 
\code{\link{orgmsCollection.init}()},
\code{\link{orgmsCollectionTestResult}}
}
\concept{orgmsCollection}
