% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgmsCollection.R
\name{orgmsCollection.init}
\alias{orgmsCollection.init}
\title{orgmsCollection.init
Initiate the orgmsCollection to an empty GRangeList}
\usage{
orgmsCollection.init(orgmsCollection)
}
\arguments{
\item{orgmsCollection}{A GRangesList object that contains a list of search targets in GRange form}
}
\value{
A empty GRangesList object
}
\description{
orgmsCollection.init
Initiate the orgmsCollection to an empty GRangeList
}
\examples{
orgmsCollection.init (orgmsCollection)
}
\seealso{
Other orgmsCollection: 
\code{\link{orgmsAdd}()},
\code{\link{orgmsCollectionTestResult}}
}
\concept{orgmsCollection}
