% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patristicDistances.R
\name{patristicToDivergenceTime}
\alias{patristicToDivergenceTime}
\title{Convert patristic distance to divergence time}
\usage{
patristicToDivergenceTime(
  patristic_distance,
  mutation_rate = 1e-09,
  generation_time = 1
)
}
\arguments{
\item{patristic_distance}{Patristic distance in substitutions per site}

\item{mutation_rate}{Mutation rate per site per year (default: 1e-9)}

\item{generation_time}{Generation time in years (default: 1)}
}
\value{
Divergence time in millions of years
}
\description{
Converts patristic distance to estimated divergence time using molecular clock
}
