% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorHandling.R
\name{provideErrorGuidance}
\alias{provideErrorGuidance}
\title{Provide specific guidance for different error types}
\usage{
provideErrorGuidance(error_type, context = list(), verbose = FALSE)
}
\arguments{
\item{error_type}{A character string specifying the type of error}

\item{context}{A list containing additional context about the error}

\item{verbose}{A logical value indicating whether to print additional information. Defaults to FALSE.}
}
\value{
A list containing guidance and solutions
}
\description{
This function provides specific guidance and solutions for different types of errors.
}
\examples{
\dontrun{
  # Get guidance for coordinate format error
  guidance <- provideErrorGuidance("coordinate_format", list(coords = "invalid"))
}

}
