% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integratedWorkflow.R
\name{suggestWorkflowModifications}
\alias{suggestWorkflowModifications}
\title{Suggest workflow modifications for incompatible parameters}
\usage{
suggestWorkflowModifications(incompatible_parameters, verbose = FALSE)
}
\arguments{
\item{incompatible_parameters}{A list containing incompatible parameters}

\item{verbose}{A logical value indicating whether to print additional information. Defaults to FALSE.}
}
\value{
A list containing suggested modifications
}
\description{
This function suggests modifications to make incompatible workflows work.
}
\examples{
\dontrun{
  # Suggest modifications for incompatible workflow
  suggestions <- suggestWorkflowModifications(
    list(species_list = c("Hsapiens", "Athaliana"), coords_source = "invalid")
  )
}

}
