% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synvizDataTest-data.R
\docType{data}
\name{synvizDataTest}
\alias{synvizDataTest}
\title{A synvizData object that is used to test the integrity of function \code{synvizData}}
\format{
A list object with \code{itrack}, \code{gtrack} and \code{atrack}
}
\source{
org.Hs.eg.db and TxDb.Hsapiens.UCSC.hg38.knownGene
}
\usage{
data(synvizDataTest)
}
\description{
synvizDataTest is a data.frame that generated by function \code{synvizData}
with given orgm = "Hsapiens" and mycoords.list is "2:16e7:16.5e7"
}
\examples{
data(synvizDataTest)
itrack <- synvizDataTest$itrack
gtrack <- synvizDataTest$gtrack
atrack <- synvizDataTest$atrack
}
\seealso{
Other SynvizPlot: 
\code{\link{multisynvizPlots}()},
\code{\link{synvizPlot}()},
\code{\link{synvizPlotData}()}
}
\concept{SynvizPlot}
\keyword{datasets}
