% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynVizOrgms-data.R
\docType{data}
\name{SynVizOrgms}
\alias{SynVizOrgms}
\title{A data table that is used to store available organisms for analysis}
\format{
A data.frame with tags: "Scientific.Names", "Common.Names", "Abbrav", "Abbrav.Short"
}
\source{
Currently available orgDB objects from
\href{http://bioconductor.org/packages/release/BiocViews.html#___AnnotationData}{bioconductor}
}
\usage{
data(SynVizOrgms)
}
\description{
SynVizOrgms is a data.frame that stores available organisms for analysis
}
\examples{
data(SynVizOrgms)
SciName <- SynVizOrgms$Scientific.Names
ComName <- SynVizOrgms$Common.Names
Abbrav <- SynVizOrgms$Abbrav
ShortAbbrav <- SynVizOrgms$Abbrav.Short
}
\concept{SyntenyViz}
\keyword{datasets}
