% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synvizPlot.R
\name{synvizPlot}
\alias{synvizPlot}
\title{synvizPlot
A plot function to plot synteny conversation across species.}
\usage{
synvizPlot(mycoords.gr, orgm)
}
\arguments{
\item{orgm}{An organism that of interest. Options include:\cr
Btaurus(Bt) - Bos taurus - Cow \cr
Celegans(Ce) - Caenorhabditis elegans - Roundworm \cr
Cfamiliaris(Cf) - Canis familiaris - Dog \cr
Dmelanogaster(Dm) - Drosophila melanogaster - Fruitfly \cr
Drerio(Dr) - Danio rerio - Zebrafish \cr
Ggallus(Gg) - Gallus gallus - Chicken \cr
Hsapiens(Hs) - Homo Sapiens - Human \cr
Mmulatta(Mmu) - Macaca mulatta - Rhesus macaque \cr
Mmusculus(Mm) - Mus musculus - House mouse \cr
Ptroglodytes(Pt) - Pan troglodytes - Chimpanzee \cr
Rnorvegicus(Rn) - Rattus norvegicus - Brown rat \cr
Scerevisiae(Sc) - Saccharomyces cerevisiae - Brewer's yeast \cr
Sscrofa(Ss) - Sus scrofa - Wild swine}
}
\value{
A synteny plot of a given gene region within a given organism
}
\description{
synvizPlot
A plot function to plot synteny conversation across species.
}
