% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSubsetTestResult-data.R
\docType{data}
\name{geneSubsetTestResult}
\alias{geneSubsetTestResult}
\title{A GRange Object}
\format{GRange}
\source{
org.Hs.eg.db and TxDb.Hsapiens.UCSC.hg38.knownGene
}
\usage{
data(geneSubsetTestResult)
}
\description{
An expected output GRange object given orgm = "Hsapiens" and mycoords.list is "2:16e7:16.5e7"
}
\examples{
data(geneSubsetTestResult)
}
\concept{geneSubset}
\keyword{datasets}
