% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synvizPlot.R
\name{multisynvizPlots}
\alias{multisynvizPlots}
\title{multisynvizPlots}
\usage{
multisynvizPlots(orgmsCollection)
}
\arguments{
\item{orgmsCollection}{A list of}
}
\description{
multisynvizPlots
}
\examples{
\dontrun {

}
}
\seealso{
Other SynvizPlot: 
\code{\link{synvizDataTest}},
\code{\link{synvizPlotData}()},
\code{\link{synvizPlot}()}
}
\concept{SynvizPlot}
