% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgmsCollection.R
\name{orgmsAdd}
\alias{orgmsAdd}
\title{orgmsAdd
Add a new search target in GRange format into GRangeList}
\usage{
orgmsAdd(orgm, orgTxDB, mycoords.list, orgmsCollection)
}
\arguments{
\item{orgm}{An organism that of interest. Options include:\cr
Btaurus(Bt) - Bos taurus - Cow \cr
Celegans(Ce) - Caenorhabditis elegans - Roundworm \cr
Cfamiliaris(Cf) - Canis familiaris - Dog \cr
Dmelanogaster(Dm) - Drosophila melanogaster - Fruitfly \cr
Drerio(Dr) - Danio rerio - Zebrafish \cr
Ggallus(Gg) - Gallus gallus - Chicken \cr
Hsapiens(Hs) - Homo Sapiens - Human \cr
Mmulatta(Mmu) - Macaca mulatta - Rhesus macaque \cr
Mmusculus(Mm) - Mus musculus - House mouse \cr
Ptroglodytes(Pt) - Pan troglodytes - Chimpanzee \cr
Rnorvegicus(Rn) - Rattus norvegicus - Brown rat \cr
Scerevisiae(Sc) - Saccharomyces cerevisiae - Brewer's yeast \cr
Sscrofa(Ss) - Sus scrofa - Wild swine}

\item{orgTxDB}{A data.frame that contains currently TxDB objects, e.g. \code{\link{orgmTxDB}}.}

\item{mycoords.list}{A string list with the format of "2:16e7:16.5e7"}

\item{orgmsCollection}{A GRangesList object that contains a list of search targets in GRange form}
}
\value{
orgmsCollection A GRangesList of search targets in GRange form
}
\description{
orgmsAdd
Add a new search target in GRange format into GRangeList
}
\examples{
orgm <- "Hsapiens"
mycoords.list <- "2:15.95e7:16.45e7"
orgm.2 <- "Mmusculus"
mycoords.list.2 <- "2:6.0e7:6.5e7"
orgm.3 <- "Rnorvegicus"
mycoords.list.3 <- "3:4.6e7:5.1e7"
\donttest{
  orgmsList <- orgmsCollection.init (orgmsList)
  orgmsList <- orgmsAdd (orgm, orgmTxDB, mycoords.list, orgmsList)
  orgmsList <- orgmsAdd (orgm.2, orgmTxDB, mycoords.list.2, orgmsList)
  orgmsList <- orgmsAdd (orgm.3, orgmTxDB, mycoords.list.3, orgmsList)
  orgmsList
}
}
\seealso{
Other orgmsCollection: 
\code{\link{orgmsCollection.init}()},
\code{\link{orgmsCollectionTestResult}}
}
\concept{orgmsCollection}
