% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgmsCollectionTestResult-data.R
\docType{data}
\name{orgmsCollectionTestResult}
\alias{orgmsCollectionTestResult}
\title{A list contains expected output for test functions.}
\format{GRangesList}
\source{
org.Hs.eg.db and TxDb.Hsapiens.UCSC.hg38.knownGene
}
\usage{
data(orgmsCollectionTestResult)
}
\description{
An expected output contains:
\describe{
  \item{add}{A GRangesList object contains info derived from \code{orgm} = "Hsapiens"
     and \code{mycoords.list} is "2:16e7:16.5e7" under tag "add"}
  \item{init}{An empty GRangesList under tag "init"}
}
}
\examples{
data(orgmsCollectionTestResult)
exp_init <- orgmsCollectionTestResult$init
exp_add  <- orgmsCollectionTestResult$add
}
\seealso{
Other orgmsCollection: 
\code{\link{orgmsAdd}()},
\code{\link{orgmsCollection.init}()}
}
\concept{orgmsCollection}
\keyword{datasets}
