% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgmTxDB-data.R
\docType{data}
\name{orgmTxDB}
\alias{orgmTxDB}
\title{A list of available TxDB objects as a data.frame}
\format{A data.frame in table format}
\source{
Currently available TxDb objects from
\href{http://bioconductor.org/packages/release/BiocViews.html#___AnnotationData}{bioconductor}
}
\usage{
data(orgmTxDB)
}
\description{
orgmTxDB is a data.frame that contains currently available TxDB objects from bioconductor
}
\examples{
data(orgmTxDB)
databaseClass <- orgmTxDB$dbClass
databaseSources <- orgmTxDB$dbSource
species <- orgmTxDB$dbSpecies
speciesAbbraviations <- orgmTxDB$dbAbbv
databaseTypes <- orgmTxDB$dbAbbv
}
\seealso{
Other orgTxDB: 
\code{\link{getTxDB}()}
}
\concept{orgTxDB}
\keyword{datasets}
