% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntenyBlocks.R
\name{plotSyntenyBlocks}
\alias{plotSyntenyBlocks}
\title{Plot synteny blocks with ortholog connections}
\usage{
plotSyntenyBlocks(
  synteny_data,
  plot_type = "comparative",
  show_connections = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{synteny_data}{A list from createSyntenyBlockData() containing synteny block data}

\item{plot_type}{A character string specifying plot type: "comparative" (side-by-side) or "overlay" (stacked)}

\item{show_connections}{A logical value indicating whether to show ortholog connections}

\item{verbose}{A logical value indicating whether to print additional information. Defaults to FALSE.}
}
\value{
A Gviz plot object
}
\description{
This function visualizes synteny blocks with ortholog connections
using Gviz tracks.
}
\examples{
\dontrun{
  # Create synteny data and plot
  synteny_data <- createSyntenyBlockData(source_coords, target_coords, 
                                        "Hsapiens", "Mmusculus")
  plotSyntenyBlocks(synteny_data, plot_type = "comparative")
}
}
