% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synvizPlot.R
\name{synvizPlotData}
\alias{synvizPlotData}
\title{synvizPlotData
A data retrive function to prepare the data for \code{synvizPlot}.}
\usage{
synvizPlotData(mycoords.gr, orgm)
}
\arguments{
\item{mycoords.gr}{A target search range in form of a GRanges object}

\item{orgm}{An organism that of interest. Options include:\cr
Btaurus(Bt) - Bos taurus - Cow \cr
Celegans(Ce) - Caenorhabditis elegans - Roundworm \cr
Cfamiliaris(Cf) - Canis familiaris - Dog \cr
Dmelanogaster(Dm) - Drosophila melanogaster - Fruitfly \cr
Drerio(Dr) - Danio rerio - Zebrafish \cr
Ggallus(Gg) - Gallus gallus - Chicken \cr
Hsapiens(Hs) - Homo Sapiens - Human \cr
Mmulatta(Mmu) - Macaca mulatta - Rhesus macaque \cr
Mmusculus(Mm) - Mus musculus - House mouse \cr
Ptroglodytes(Pt) - Pan troglodytes - Chimpanzee \cr
Rnorvegicus(Rn) - Rattus norvegicus - Brown rat \cr
Scerevisiae(Sc) - Saccharomyces cerevisiae - Brewer's yeast \cr
Sscrofa(Ss) - Sus scrofa - Wild swine}
}
\value{
A list that consists of AnnotationTrack, GenomeAxisTrack and IdeogramTrack
to feed into the function \code{synvizPlot}
}
\description{
synvizPlotData
A data retrive function to prepare the data for \code{synvizPlot}.
}
\examples{
\dontrun{
  orgm <- "Hsapiens"
  mycoords.list <- "2:16e7:16.5e7"
  mycoords.gr <- coordFormat (mycoords.list)
  synvizPlotData (mycoords.gr, orgm)
}
}
\seealso{
Other SynvizPlot: 
\code{\link{synvizDataTest}},
\code{\link{synvizPlot}()}
}
\concept{SynvizPlot}
